# Conflicts:
#	src/main/java/com/microsoft/sqlserver/jdbc/IOBuffer.java

+++ src/main/java/com/microsoft/sqlserver/jdbc/Column.java
@@ -201,7 +201,7 @@ final class Column {
     }
 
     int getInt(TDSReader tdsReader) throws SQLServerException {
-        return ((Integer) getValue(JDBCType.INTEGER, null, null, tdsReader)).intValue();
+        return (Integer) getValue(JDBCType.INTEGER, null, null, tdsReader);
     }
 
     void updateValue(JDBCType jdbcType,
